+++ History.md
@@ -1,5 +1,6 @@
 ### 0.7.0 ###
 - ReSharper 2017.1 support
+- Added provider for "Copy Fully-qualified name" which returns <TypeShortName>.<MemberShortName>
 
 ### 0.6.0 ###
 - ReSharper 2016.3 support
@@ -17,4 +18,5 @@
 - ReSharper 9.2 support
 
 ### 0.1 ###
+- Added Live Template Macro which returns the current Git branch
 - ReSharper 9.1 support